refactor(phase3b): model I/O・hash・metadata 系を library/model_io.py に分離
train_util.py から以下を library/model_io.py に移動:
- ハッシュ系: model_hash / calculate_sha256 / addnet_hash_legacy /
  addnet_hash_safetensors / precalculate_safetensors_hashes / get_git_revision_hash
- attention 切替: replace_unet_modules
- safetensors metadata: load_metadata_from_safetensors
- ss_* metadata: SS_METADATA_KEY_* 定数 + SS_METADATA_MINIMUM_KEYS +
  build_minimum_network_metadata
- SAI ModelSpec: get_sai_model_spec / get_sai_model_spec_dataclass
- model loader: _load_target_model / load_target_model

計 13 関数 + 7 定数。train_util.py は re-export shim として維持。
networks/lora_interrogator.py が train_util._load_target_model を参照しているため
shim にも _load_target_model を含める。

train_util.py: 2,449 → 2,103 行
library/model_io.py: 421 行（新規）
